Remove hide_window support
This never worked correctly for me anyway (at least in Gnome 3.14), and
we need a module to substitute python-keybinder2 in PyGObject (since
python-keybinder2 only works on PyGTK).

The actual function that seems to actually hides the window
(terminatorlib/window.py/on_hide_window()) is still there, though. So if
there is another way to bind hide_window to a (global?) shortcut, we can
implement it.